Move tests to standalone module
